+++ NEWS.md
@@ -1,5 +1,7 @@
 # glyrepr (development version)
 
+* Add `as.list()` support for `glyrepr_composition` and `glyrepr_structure` vectors. (#4)
+
 # glyrepr 0.12.1
 
 ## Minor improvements and bug fixes